+++ common/model/src/main/java/is/codion/common/model/condition/ConditionModel.java
@@ -451,6 +451,7 @@ public interface ConditionModel<T> {
 		Builder<T> caseSensitive(boolean caseSensitive);
 
 		/**
+		 * Default true.
 		 * @param autoEnable true if the model should auto-enable
 		 * @return this builder instance
 		 */